Release/1.7.4 (#257)
* Add a progress bar navigation to the dating wizard

* Add a link to dating preview screen to dating settings page

* Add tests

* Fix eslint issues

* Fix navigation issues

* Show specific error message for dating enable request

* Add UID to graphQL headers

* Add tests for postsSingleGet reducer

* Add loading state

* Fix copies

* Add tests

* Fix edit profile form (#242)

* [Trello 202] Find friends as an empty state for the feed (#236)

* Replace cannot to can`t

* Find friends as an empty state for the feed

* [Trello 306] Fix Sentry errors (#237)

* Prevent usersReportScreenViewsRequest for a not authorized user

* Prevent call Logger.captureException when a user cancelled image selection

* Prevent Setry capture errors for cancelled by signout requests

* Fix an error null is not an object (evaluating 'C.current.scrollToOffset')

* Prevent postsReportPostViewsRequest for not authorized user

* Prevent FEED_CARDS_UNSUPPORTED_CARD error

* Fix unused takenInReal variable error

* Add sign in button to profile upgrade screen (#238)

* Fix post report (#239)

* [Trello 114] Prevent choosing multiply photos from library for Profile photo (#241)

* Prevent choosing multiply photos from library for Profile photo

* Prevent choosing multiply photos from library for Profile photo

* [Trello 94] Set the TextInput touched only on blur (#223)

* Set the TextInput touched only on blur

* Removed autoFocus

* Dating display name

* Fix copies (#245)

* Always refresh feed each time when app open (#246)

* Clear comment form on success (#247)

* Height select initialValue should be in a middle of range (#244)

* Fix subscription (#248)

* Fix infinity loading state (#249)

* Rename "Post" button in chat to say "Send" (#251)

* [Trello 314] Snackbars (#253)

* Snackbars

* Update blacklist for snackbars

* Update podlock file

* Remove console.log

* Fix the black profile issue after about 7 days of using the app (#254)

* [Trello 327] Make username editable (#255)

* Make username editable

* Username field attributes

* Diamond Trial subscription (#252)

* [Trello-356] Update EULA and ToS links (#256)

* Testflight artifacts

+++ src/components/Comments/index.service.js
@@ -27,6 +27,19 @@ const CommentsService = ({ children }) => {
   const postsSingleGet = useSelector(postsSelector.postsSingleGetSelector(postId))
 
   const commentsRef = useRef()
+  const formRef = useRef()
+
+  const resetForm = () => {
+    try {
+      formRef.current.resetForm()
+      
+      setTimeout(() => {
+        Keyboard.dismiss()
+      }, 0)
+    } catch(error) {
+      // ignore
+    }
+  }
 
   useEffect(() => {
     const commentIndex = postsCommentsGet.data.findIndex(item => item.commentId === actionId)
@@ -42,22 +55,29 @@ const CommentsService = ({ children }) => {
   }, [])
 
   useEffect(() => {
-    dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.commentsAddIdle({}))
-  }, [commentsAdd.status === 'success'])
+    if (commentsAdd.status === 'success') {
+      dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.commentsAddIdle({}))
+      resetForm()
+    }
+  }, [commentsAdd.status])
 
   useEffect(() => {
-    dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.commentsDeleteIdle({}))
-  }, [commentsDelete.status === 'success'])
+    if (commentsDelete.status === 'success') {
+      dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.commentsDeleteIdle({}))
+    }
+  }, [commentsDelete.status])
 
   useEffect(() => {
-    dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
-    dispatch(postsActions.commentsFlagIdle({}))
-  }, [commentsFlag.status === 'success'])
+    if (commentsFlag.status === 'success') {
+      dispatch(postsActions.postsCommentsGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.postsSingleGetRequest({ postId, userId: postUserId }))
+      dispatch(postsActions.commentsFlagIdle({}))
+    }
+  }, [commentsFlag.status])
 
   const commentsAddRequest = ({ text }) => {
     const commentId = uuid()
@@ -99,8 +119,6 @@ const CommentsService = ({ children }) => {
 
   const marginBottom = offset + ifIphoneX(40, 0)
   
-  
-
   /**
    * FlatList feed config ref, used for reporting scroll events
    */
@@ -117,41 +135,32 @@ const CommentsService = ({ children }) => {
     text: useRef(null),
   })
 
-  const [replyUser, setReplyUser] = useState(null)
   const handleUserReply = (username) => {
-    setReplyUser(`@${username} `)
-    if (inputRefs.current.text.current) {
+    try {
+      formRef.current.setFieldValue('text', `@${username} `)
       inputRefs.current.text.current.focus()
+    } catch(error) {
+      // ignore
     }
   }
 
-  const handleFormSubmit = (values, { resetForm }) => {
+  const handleFormSubmit = (values) => {
     commentsAddRequest(values)
-    resetForm()
-    Keyboard.dismiss()
   }
 
   const formSubmitLoading = commentsAdd.status === 'loading'
   const formSubmitDisabled = commentsAdd.status === 'loading'
-  const formErrorMessage = commentsAdd.error.text
 
   const formInitialValues = {
-    text: replyUser,
+    text: '',
   }
 
   const handleFormTransform = (values) => ({
     text: compose(trim, pathOr('', ['text']))(values),
   })
 
-  const handleErrorClose = () => dispatch(postsActions.commentsAddIdle({}))
-
-
-  
-
   return children({
     user,
-    commentsAdd,
-    commentsAddRequest,
     commentsDeleteRequest,
     commentsFlagRequest,
     postsCommentsGet,
@@ -162,14 +171,13 @@ const CommentsService = ({ children }) => {
     handleUserReply,
     commentsRef,
 
-    formErrorMessage,
     handleFormSubmit,
     handleFormTransform,
-    handleErrorClose,
     formSubmitLoading,
     formSubmitDisabled,
     formInitialValues,
     inputRefs,
+    formRef,
   })
 }
